# public/swagger.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"Users found (can be empty)","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}}},"tags":["user"]}},"/users/unit/{id}":{"get":{"operationId":"UserController_findByUnitId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Users found (can be empty)","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"tags":["user"]}},"/user/create/event":{"post":{"operationId":"UserController_createEvent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"Event created","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/Event"}}}}}}},"tags":["user"],"security":[{"bearer-token":[]}]}},"/user/join/event/{id}":{"post":{"operationId":"UserController_joinEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User joined event","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"object","$ref":"#/components/schemas/Participant"}}}}}}},"tags":["user"],"security":[{"bearer-token":[]}]}},"/events":{"get":{"operationId":"EventController_findAll","parameters":[],"responses":{"200":{"description":"List of events (can be empty)","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Event"}}}}}}}},"tags":["event"]}},"/event/{id}":{"get":{"operationId":"EventController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Event found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}}},"tags":["event"]}},"/events/unit/{id}":{"get":{"operationId":"EventController_findByUnitId","summary":"Find all events for a unit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Event found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}}},"tags":["event"]}},"/units":{"get":{"operationId":"UnitController_findAll","parameters":[],"responses":{"200":{"description":"List of units (can be empty)","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Unit"}}}}}}}},"tags":["unit"]}},"/unit/{id}":{"get":{"operationId":"UnitController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Unit found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Unit"}}}}}}}},"tags":["unit"]}},"/units/user/{id}":{"get":{"operationId":"UnitController_findByUserId","summary":"Find all units for a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Unit found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Unit"}}}}}}}},"tags":["unit"]}},"/units/event/{id}":{"get":{"operationId":"UnitController_findByEventId","summary":"Find all units for an event","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Unit found","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Unit"}}}}}}}},"tags":["unit"]}},"/unit":{"post":{"operationId":"UnitController_create","parameters":[],"responses":{"200":{"description":"Unit created","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/Unit"}}}}}}},"tags":["unit"]}},"/event-units":{"get":{"operationId":"EventUnitController_findAll","summary":"Getting the Unit&Event relation table.","parameters":[],"responses":{"200":{"description":"Get all event units","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/EventUnit"}}}}}}}},"tags":["event-unit"]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"User created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"}}}}}}}},"400":{"description":"User already exist. Please login.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"error":{"type":"string"}}}}}}},"tags":["auth"]}},"/auth/admin/signup":{"post":{"operationId":"AuthController_adminSignUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"User created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"admin":{"$ref":"#/components/schemas/Admin"}}}}}}}}},"tags":["auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User signed in successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"object","properties":{"token":{"type":"string"}}}}}}}}},"tags":["auth"]}},"/auth/protected":{"get":{"operationId":"AuthController_protected","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer-token":[]}]}},"/auth/signout":{"get":{"operationId":"AuthController_signOut","parameters":[],"responses":{"200":{"description":"User signed out successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","example":"Signout complete."}}}}}}},"tags":["auth"],"security":[{"bearer-token":[]}]}},"/admins":{"get":{"operationId":"AdminController_findAll","parameters":[],"responses":{"200":{"description":"List of all admins.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Admin"}}}}}}}},"tags":["admin"]}},"/admin/event/{id}/status":{"put":{"operationId":"AdminController_manageEvent","summary":"Approve or reject an event.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventStatusDto"}}}},"responses":{"200":{"description":"Event status updated successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}}}}}}},"tags":["admin"],"security":[{"bearer-token":[]}]}},"/unit-admins":{"get":{"operationId":"UnitAdminController_findAll","summary":"Getting the Unit&Admin relation table.","parameters":[],"responses":{"200":{"description":"Get all unit admins","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/UnitAdmin"}}}}}}}},"tags":["unit-admin"]}},"/unit-admins/unit/{id}":{"get":{"operationId":"UnitAdminController_findByUnitId","summary":"Getting Admins by unit ID.","parameters":[{"name":"id","required":true,"in":"path","description":"Unit ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all admins by unit ID","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/UnitAdmin"}}}}}}}},"tags":["unit-admin"]}},"/unit-admins/admin/{id}":{"get":{"operationId":"UnitAdminController_findByAdminId","summary":"Getting Units by admin ID.","parameters":[{"name":"id","required":true,"in":"path","description":"Admin ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all units by admin ID","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/UnitAdmin"}}}}}}}},"tags":["unit-admin"]}},"/orgs":{"get":{"operationId":"OrganizerController_findAll","parameters":[],"responses":{"200":{"description":"Get all organizers","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"array","items":{"$ref":"#/components/schemas/Organizer"}}}}}}}},"tags":["organizer"]}},"/org/{id}":{"get":{"operationId":"OrganizerController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get organizer by ID","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string","$ref":"#/components/schemas/Organizer"}}}}}}},"tags":["organizer"]}},"/org/event/{eventId}/basic-info":{"put":{"operationId":"OrganizerController_update","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventDto"}}}},"responses":{"200":{"description":""}},"tags":["organizer"],"security":[{"bearer-token":[]}]}},"/org/event/{eventId}/status":{"put":{"operationId":"OrganizerController_updateStatus","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventStatusDto"}}}},"responses":{"200":{"description":""}},"tags":["organizer"],"security":[{"bearer-token":[]}]}},"/org/event/{eventId}/participants":{"put":{"operationId":"OrganizerController_manageParticipant","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateParticipantStatusDto"}}}},"responses":{"200":{"description":""}},"tags":["organizer"],"security":[{"bearer-token":[]}]}},"/org/event/{eventId}/unit":{"put":{"operationId":"OrganizerController_manageEventUnit","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateEventUnitDto"}}}},"responses":{"200":{"description":""}},"tags":["organizer"],"security":[{"bearer-token":[]}]}},"/participants":{"get":{"operationId":"ParticipantController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["participant"]}},"/participant/{id}":{"get":{"operationId":"ParticipantController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["participant"]}},"/participant/leave/event/{id}":{"put":{"operationId":"ParticipantController_leaveEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["participant"],"security":[{"bearer-token":[]}]}}},"info":{"title":"Event Management API","description":"The Event Management API description","version":"1.0","contact":{}},"tags":[{"name":"swagger","description":""}],"servers":[],"components":{"securitySchemes":{"bearer-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","description":"Enter JWT token","in":"header"}},"schemas":{"User":{"type":"object","properties":{"fId":{"type":"string","example":"<firebase id>","description":"Firebase Id obtained from Google Auth"},"firstName":{"type":"string","example":"Updated","description":"Updated"},"lastName":{"type":"string","example":"Doe","description":"Last Name"},"email":{"type":"string","example":"JohnDoe@gmail.com","description":"Email Address"},"gender":{"type":"string","example":"male","description":"Gender"},"age":{"type":"number","example":25,"description":"Age"},"phone":{"type":"number","example":1234567890,"description":"Phone Number"},"isAdmin":{"type":"boolean","example":"false","description":"Is Admin Boolean Value"}},"required":["fId","firstName","lastName","email","gender","age","phone","isAdmin"]},"CreateEventDto":{"type":"object","properties":{"name":{"type":"string","example":"Event Name","description":"Name of the event"},"unitId":{"type":"string","example":"<mongoose _id>","description":"Unit ID"}},"required":["name","unitId"]},"CreateParticipantDto":{"type":"object","properties":{"userId":{"type":"string","example":"<mongoose _id>","description":"The ID of the user"},"status":{"type":"string","example":"pending | participating | participated | rejected","description":"One of the Status."},"organizerId":{"type":"string","example":"<mongoose _id>","description":"The ID of the Org."}},"required":["userId","status","organizerId"]},"Event":{"type":"object","properties":{"name":{"type":"string","example":"Event name","description":"Name of the event"},"description":{"type":"string","example":"Event description","description":"Description of the event"},"startDate":{"type":"string","example":"2022-12-31","description":"Start date of the event"},"endDate":{"type":"string","example":"2022-12-31","description":"End date of the event"},"startTime":{"type":"string","example":"08:00","description":"Start time of the event"},"endTime":{"type":"string","example":"17:00","description":"End time of the event"},"location":{"type":"string","example":"John Paul II","description":"Location of the event"},"rules":{"type":"string","example":"No alcohol","description":"Rules of the event"},"coverImageUrL":{"type":"string","example":"Event cover image URL","description":"Cover image URL of the event"},"adminId":{"type":"string","example":"<mongoose _id>","description":"Admin ID"},"status":{"type":"string","example":"pending | approved | active | suspended | cancelled | completed | rejected","description":"Status of the event"}},"required":["name","description","startDate","endDate","startTime","endTime","location","rules","coverImageUrL","adminId","status"]},"Organizer":{"type":"object","properties":{"userId":{"example":"<userId>","description":"User ID","allOf":[{"$ref":"#/components/schemas/User"}]},"eventId":{"example":"<eventId>","description":"Event ID","allOf":[{"$ref":"#/components/schemas/Event"}]},"status":{"type":"string","example":"organizing","description":"Status [ORGANIZING, ORGANIZED]"}},"required":["userId","eventId","status"]},"UpdateEventDto":{"type":"object","properties":{"name":{"type":"string","example":"Event Name","description":"Name of the event"},"organizer":{"example":"Event Description","description":"Description of the event","allOf":[{"$ref":"#/components/schemas/Organizer"}]},"status":{"type":"string","description":"Don't pass status. Use update-event-status api. Ask developer for more info."}},"required":["name","organizer","status"]},"UpdateEventStatusDto":{"type":"object","properties":{"status":{"type":"string","example":"pending | approved | active | suspended | cancelled | completed | rejected","description":"One of the Status of the event."},"adminId":{"type":"string","example":"<admin id>","description":"Id of the admin who is updating the status."}},"required":["status","adminId"]},"Unit":{"type":"object","properties":{"name":{"type":"string","example":"Unit Name","description":"Name of the unit"},"description":{"type":"string","example":"Unit Description","description":"Description of the unit"}},"required":["name","description"]},"UnitMember":{"type":"object","properties":{"userId":{"example":"<mongoose _id>","description":"User ID","allOf":[{"$ref":"#/components/schemas/User"}]},"unitId":{"example":"<mongoose _id>","description":"Unit ID","allOf":[{"$ref":"#/components/schemas/Unit"}]}},"required":["userId","unitId"]},"EventUnit":{"type":"object","properties":{"eventId":{"example":"<eventId>","description":"Event ID","allOf":[{"$ref":"#/components/schemas/Event"}]},"unitId":{"example":"<unitId>","description":"Unit ID","allOf":[{"$ref":"#/components/schemas/Unit"}]}},"required":["eventId","unitId"]},"SignUpDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John","description":"First Name"},"email":{"type":"string","example":"Doe","description":"Last Name"},"phone":{"type":"number","example":"123456788","description":"Phone Number"},"fId":{"type":"string","example":"<firebase id>","description":"Firebase Id obtained from Google Auth"},"unitId":{"type":"string","example":"<mongoose _id>","description":"Unit ID"}},"required":["firstName","email","phone","fId","unitId"]},"SignInDto":{"type":"object","properties":{"fId":{"type":"string","example":"<firebase id>","description":"Firebase Id obtained from Google Auth"}},"required":["fId"]},"RestrictedToken":{"type":"object","properties":{"token":{"type":"string","example":"<jwt token>","description":"Token obtained from SignOut"}},"required":["token"]},"SignOutDto":{"type":"object","properties":{"token":{"type":"string","example":"<jwt token>","description":"Token obtained from SignIn"}},"required":["token"]},"Admin":{"type":"object","properties":{"userId":{"example":"<userId>","description":"User ID","allOf":[{"$ref":"#/components/schemas/User"}]},"unitId":{"example":"<unitId>","description":"Unit ID","allOf":[{"$ref":"#/components/schemas/Unit"}]}},"required":["userId","unitId"]},"UnitAdmin":{"type":"object","properties":{"adminId":{"example":"adminId","description":"Admin ID","allOf":[{"$ref":"#/components/schemas/Admin"}]},"unitId":{"example":"unitId","description":"Unit ID","allOf":[{"$ref":"#/components/schemas/Unit"}]}},"required":["adminId","unitId"]},"UpdateParticipantStatusDto":{"type":"object","properties":{"participants":{"example":["<mongoose _id>"],"description":"The ID List of the participants","type":"array","items":{"type":"string"}},"status":{"type":"string","example":"accepted | rejected | kicked","description":"One of the Status."}},"required":["participants","status"]},"updateEventUnitDto":{"type":"object","properties":{"action":{"type":"string","example":"remove","description":"Action to perform"},"units":{"example":["<unitId>"],"description":"Unit IDs","type":"array","items":{"type":"string"}}},"required":["action","units"]},"Participant":{"type":"object","properties":{"userId":{"example":"<mongoose _id>","description":"User ID","allOf":[{"$ref":"#/components/schemas/User"}]},"eventId":{"example":"<mongoose _id>","description":"Event ID","allOf":[{"$ref":"#/components/schemas/Event"}]},"organzierId":{"example":"<mongoose _id>","description":"Organizer ID","allOf":[{"$ref":"#/components/schemas/Organizer"}]},"status":{"type":"string","example":"pending | participating | participated | rejected","description":"Status of the participant"},"email":{"type":"string","example":"JohnDoe@gmail.com","description":"Email of the participant"},"phone":{"type":"number","example":"123456789","description":"Phone Number"}},"required":["userId","eventId","organzierId","status","email","phone"]}}}}